Working indexed tree implementation with graphite glob searches
